refactor: move app logic to app module, update os --cpus output
